Change gromacs.py back to using spack only
